backend/requirements.txt
asgiref==3.3.1
astroid==2.4.2
Django==3.1.5
django-extensions==3.0.9
djangorestframework==3.12.2
Faker==5.4.0
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
wrapt==1.12.1
yapf==0.30.0
